mentor styles merging
