+++ client/src/Components/Mentors/MentorsHome/MentorsGuide.styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const MentorsGuideStyles = styled.div`
+	width: 100%;
+	height: 100%;
+	overflow: hidden;
+	display: flex;
+	flex-flow: row wrap;
+`;
