+++ client/src/pages/registration.styles.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const Wrapper = styled.div`
+export const RegistrationStyles = styled.div`
 	width: 100%;
 	height: 100vh;
